+++ build.gradle
@@ -21,7 +21,21 @@ dependencies {
 
 test {
     useTestNG() {
-        suites "src/test/resources/testNG.xml"
+        suites "src/test/resources/TestNG.xml"
+        systemProperty "browser", System.getProperty("browser")
+    }
+}
+
+task testSmoke (type: Test) {
+    useTestNG() {
+        suites "src/test/resources/Smoke.xml"
+        systemProperty "browser", System.getProperty("browser")
+    }
+}
+
+task testRegression (type: Test) {
+    useTestNG() {
+        suites "src/test/resources/Regression.xml"
         systemProperty "browser", System.getProperty("browser")
     }
 }
