--- src/test/java/Homework18.java
@@ -1,27 +1,0 @@
-import com.beust.ah.A;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import pages.AllSongsPage;
-import pages.HomePage;
-import pages.LoginPage;
-
-public class Homework18 extends BaseTest {
-
-    @Test
-    public void playSong() {
-
-        LoginPage loginPage = new LoginPage(driver);
-        HomePage homePage = new HomePage(driver);
-        AllSongsPage allSongsPage = new AllSongsPage(driver);
-
-        loginPage.provideEmail("hand923@gmail.com")
-                 .providePassword("te$t$tudent")
-                 .clickSubmitButton();
-
-        homePage.clickAllSongs();
-
-        allSongsPage.selectFirstSong()
-                    .playSelectedSong();
-        Assert.assertTrue(allSongsPage.isPlaying().isDisplayed());
-    }
-}
