--- src/test/resources/Homework18.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="Homework18">
-    <parameter name="BaseURL" value="http://bbb.testpro.io" />
-    <test name="Example Test" preserve-order="false">
-        <classes>
-            <class name="Homework18"/>
-            <class name="BaseTest"/>
-        </classes>
-    </test>
-</suite>
