--- src/test/resources/Homework21.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="Homework21">
-    <parameter name="BaseURL" value="http://bbb.testpro.io" />
-    <test name="Example Test" preserve-order="false">
-        <classes>
-            <class name="Homework21"/>
-            <class name="BaseTest"/>
-        </classes>
-    </test>
-</suite>
