+++ src/test/resources/Regression.xml
@@ -1,11 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG" parallel="methods" thread-count="1">
+<suite name="Regression" parallel="methods" thread-count="1">
+    <groups>
+        <run>
+            <include name="regression"/>
+        </run>
+    </groups>
+
     <parameter name="BaseURL" value="http://bbb.testpro.io" />
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="BaseTest"/>
             <class name="LoginTests"/>
             <class name="CustomPlaylistTests"/>
             <class name="PlaySongsTests"/>
